Change class name and move it to more proper package
