+++ src/model/chess_pieces/King.java
@@ -2,7 +2,6 @@ package model.chess_pieces;
 
 import model.Board;
 import model.BoardPosition;
-import model.MovementManager;
 import model.Player;
 
 import java.util.LinkedList;
@@ -17,8 +16,8 @@ public class King extends ChessPiece {
     public List<BoardPosition> getAllPossibleMovements(Board board, BoardPosition place) {
         List<BoardPosition> moves = new LinkedList<>();
 
-        moves.addAll(MovementManager.getAllForSides(    board, place, 1));
-        moves.addAll(MovementManager.getAllForDiagonals(board, place, 1));
+        moves.addAll(ChessPieceMovements.getAllForSides(    board, place, 1));
+        moves.addAll(ChessPieceMovements.getAllForDiagonals(board, place, 1));
 
         return moves;
     }
