+++ src/model/chess_pieces/Pawn.java
@@ -2,7 +2,6 @@ package model.chess_pieces;
 
 import model.Board;
 import model.BoardPosition;
-import model.MovementManager;
 import model.Player;
 
 import java.util.List;
@@ -18,6 +17,6 @@ public class Pawn extends ChessPiece {
         int direction = (getColor() == ChessPieceColor.BLACK) ? 1 : -1;
         int max = (place.getY() == 1 || place.getY() == 6) ? 2 : 1;
 
-        return MovementManager.getAllForPawn(board, place, max, direction);
+        return ChessPieceMovements.getAllForPawn(board, place, max, direction);
     }
 }
