+++ src/model/chess_pieces/Queen.java
@@ -2,7 +2,6 @@ package model.chess_pieces;
 
 import model.Board;
 import model.BoardPosition;
-import model.MovementManager;
 import model.Player;
 
 import java.util.LinkedList;
@@ -17,8 +16,8 @@ public class Queen extends ChessPiece {
     public List<BoardPosition> getAllPossibleMovements(Board board, BoardPosition place) {
         List<BoardPosition> moves = new LinkedList<>();
 
-        moves.addAll(MovementManager.getAllForSides(board, place, Integer.MAX_VALUE));
-        moves.addAll(MovementManager.getAllForDiagonals(board, place, Integer.MAX_VALUE));
+        moves.addAll(ChessPieceMovements.getAllForSides(board, place, Integer.MAX_VALUE));
+        moves.addAll(ChessPieceMovements.getAllForDiagonals(board, place, Integer.MAX_VALUE));
 
         return moves;
     }
